requirements.txt
Django==4.2.9
django-ckeditor==6.7.0
django-solo==2.2.0
Pillow==10.2.0